--- a/src/Microsoft.AspNetCore.SignalR.Core/DefaultHubActivator.cs
+++ b/src/Microsoft.AspNetCore.SignalR.Core/DefaultHubActivator.cs
@@ -9,6 +9,8 @@ namespace Microsoft.AspNetCore.SignalR
 {
     public class DefaultHubActivator<THub> : IHubActivator<THub> where THub: Hub
     {
+        // Object factory for all this Hub type
+        private static ObjectFactory _objectFactory = ActivatorUtilities.CreateFactory(typeof(THub), Type.EmptyTypes);
         private readonly IServiceProvider _serviceProvider;
         private bool? _created;
 
@@ -25,7 +27,7 @@ namespace Microsoft.AspNetCore.SignalR
             var hub = _serviceProvider.GetService<THub>();
             if (hub == null)
             {
-                hub = ActivatorUtilities.CreateInstance<THub>(_serviceProvider);
+                hub = (THub)_objectFactory(_serviceProvider, Array.Empty<object>());
                 _created = true;
             }