--- a/src/Microsoft.AspNetCore.SignalR.Core/DefaultHubActivator.cs
+++ b/src/Microsoft.AspNetCore.SignalR.Core/DefaultHubActivator.cs
@@ -10,7 +10,7 @@ namespace Microsoft.AspNetCore.SignalR
     public class DefaultHubActivator<THub> : IHubActivator<THub> where THub: Hub
     {
         // Object factory for THub instances
-        private static Lazy<ObjectFactory> _objectFactory = new Lazy<ObjectFactory>(() => ActivatorUtilities.CreateFactory(typeof(THub), Type.EmptyTypes));
+        private static readonly Lazy<ObjectFactory> _objectFactory = new Lazy<ObjectFactory>(() => ActivatorUtilities.CreateFactory(typeof(THub), Type.EmptyTypes));
         private readonly IServiceProvider _serviceProvider;
         private bool? _created;