- Turns out CreateFactory is much much faster
- Added a benchmark for hub activation

## Before

```
 Method |     Mean |    Error |   StdDev |        Op/s |  Gen 0 | Allocated |
------- |---------:|---------:|---------:|------------:|-------:|----------:|
 Create | 710.7 ns | 14.40 ns | 39.89 ns | 1,406,988.1 | 0.0153 |     224 B |
```

## After

```
  Method |     Mean |    Error |   StdDev |         Op/s |  Gen 0 | Allocated |
------- |---------:|---------:|---------:|-------------:|-------:|----------:|
 Create | 76.34 ns | 3.347 ns | 9.386 ns | 13,099,035.7 | 0.0037 |      48 B |
```

## Lazy

```
 Method |     Mean |    Error |   StdDev |   Median |         Op/s |  Gen 0 | Allocated |
------- |---------:|---------:|---------:|---------:|-------------:|-------:|----------:|
 Create | 65.39 ns | 1.480 ns | 4.102 ns | 63.73 ns | 15,292,741.1 | 0.0037 |      48 B |
```

Fixes #1570