@pakrym 2018-03-19
Why did you need full on `Lazy<T>`? Null check would be just fine.