@davidfowl 2018-03-19
> Why did you need full on Lazy<T>? Null check would be just fine.

You're right. I guess worst case we would have compiled a few times. Meh.